refactor(dispatch): group regressions by domain
